requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
frozenlist==1.8.0
googletrans==4.0.2
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
